[ci skip] Refactoring.
