+++ src/main/java/com/graffitab/server/api/authentication/JsonLoginSuccessHandler.java
@@ -16,7 +16,7 @@ import com.graffitab.server.api.dto.user.GetUserResult;
 import com.graffitab.server.api.dto.user.UserDto;
 import com.graffitab.server.api.mapper.OrikaMapper;
 import com.graffitab.server.persistence.model.User;
-import com.graffitab.server.service.UserSessionService;
+import com.graffitab.server.service.user.UserSessionService;
 
 public class JsonLoginSuccessHandler implements AuthenticationSuccessHandler {
