+++ src/main/java/com/graffitab/server/service/DeviceService.java
@@ -11,6 +11,7 @@ import com.graffitab.server.api.errors.ResultCode;
 import com.graffitab.server.persistence.dao.HibernateDaoImpl;
 import com.graffitab.server.persistence.model.Device;
 import com.graffitab.server.persistence.model.Device.OSType;
+import com.graffitab.server.service.user.UserService;
 import com.graffitab.server.persistence.model.User;
 
 @Service
